// src/models/step.enum.ts
export enum Step {
    Jira = 'jira',
    Type = 'type',
    Scope = 'scope',
    Subject = 'subject',
    Body = 'body',
    IsBreaking = 'isBreaking',
    BreakingBody = 'breakingBody',
    Breaking = 'breaking',
    IsIssueAffected = 'isIssueAffected',
    IssueBody = 'issueBody',
    IssueReference = 'issueReference',
}